Remove lynas.dev
