--- lynas_dev/static/index.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-    <title>lynas.dev</title>
-</head>
-<body>
-    <img src="images/construction.jpg">
-</body>
-</html>
-
