+++ main_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"go-app/polytopia"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -13,7 +12,7 @@ func TestPolytopiaGetGood(t *testing.T) {
         t.Fatal(err)
     }
     rr := httptest.NewRecorder()
-    handler := http.HandlerFunc(polytopia.PolytopiaHandler)
+    handler := http.HandlerFunc(polytopiaHandler)
     handler.ServeHTTP(rr, req)
     if status := rr.Code; status != http.StatusOK {
         t.Errorf("handler returned wrong status code: got %v want %v",
@@ -27,7 +26,7 @@ func TestPolytopiaBadMethod(t *testing.T) {
         t.Fatal(err)
     }
     rr := httptest.NewRecorder()
-    handler := http.HandlerFunc(polytopia.PolytopiaHandler)
+    handler := http.HandlerFunc(polytopiaHandler)
     handler.ServeHTTP(rr, req)
     if status := rr.Code; status != http.StatusMethodNotAllowed {
         t.Errorf("handler returned wrong status code: got %v want %v",
